# backend/docs/swagger.yaml
basePath: /admin-api
definitions:
  github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: ""
        type: string
    type: object
...
info:
  contact: {}
  description: 'Go modular-monolith scaffold for operations platforms. Default database:
    MySQL 8.4.'
  title: Nimbus Framework Go API
  version: "1.0"
paths:
...
  /ai-platform/integrations/{instanceId}/revisions:
    get:
      parameters:
      - description: Integration instance ID
        in: path
        name: instanceId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: List immutable configuration backup, smoke and rollback revisions
      tags:
      - VisionAI Operations
...
  /ai-platform/integrations/{instanceId}/upgrade:
    post:
      parameters:
      - description: Integration instance ID
        in: path
        name: instanceId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Upgrade an integration with compatibility gate, smoke test and automatic
        rollback
      tags:
      - VisionAI Operations
...
  /ai-platform/sync-incidents/{incidentId}/action:
    post:
      parameters:
      - description: Sync incident ID
        in: path
        name: incidentId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Replay, rebind, ignore or manually close a synchronization incident
      tags:
      - VisionAI Operations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"